comment out all of update.py
