+++ backend/ee/danswer/background/celery/celery_app.py
@@ -1,8 +1,8 @@
 from datetime import timedelta
 
 from danswer.background.celery.celery_app import celery_app
+from danswer.background.celery.celery_utils import get_all_tenant_ids
 from danswer.background.task_utils import build_celery_task_wrapper
-from danswer.background.update import get_all_tenant_ids
 from danswer.configs.app_configs import JOB_TIMEOUT
 from danswer.configs.app_configs import MULTI_TENANT
 from danswer.db.chat import delete_chat_sessions_older_than
